+++ packages/react-charts/src/components/ChartPie/ChartPie.tsx
@@ -651,7 +651,7 @@ export const ChartPie: React.FunctionComponent<ChartPieProps> = ({
   // Callback to compliment legendAllowWrap
   const computedLegend = getLegend();
   useEffect(() => {
-    if (typeof legendAllowWrap === 'function') {
+    if (computedLegend?.props && typeof legendAllowWrap === 'function') {
       const extraHeight = getLegendItemsExtraHeight({
         legendData: computedLegend.props.data,
         legendOrientation: computedLegend.props.orientation,