+++ packages/react-core/src/demos/examples/Nav/NavDrilldown.tsx
@@ -12,9 +12,7 @@ import {
   Menu,
   PageSection
 } from '@patternfly/react-core';
-
-import DashboardHeader from '@patternfly/react-core/src/demos/examples/DashboardHeader';
-
+import { DashboardHeader } from '@patternfly/react-core/src/demos/DashboardHeader';
 interface MenuHeights {
   [menuId: string]: number;
 }