+++ packages/react-table/src/deprecated/components/Table/examples/LegacyTableMisc.tsx
@@ -29,14 +29,13 @@ export const LegacyTableMisc: React.FunctionComponent = () => {
     repo.lastCommit
   ]);
 
-  const customRowWrapper: TableProps['rowWrapper'] = ({ trRef, className, rowProps, row: _row, ...props }) => {
+  const customRowWrapper: TableProps['rowWrapper'] = ({ trRef, className, rowProps, row: _row }) => {
     const isOddRow = rowProps ? !!((rowProps.rowIndex + 1) % 2) : true;
     const customStyle = {
       borderLeft: `3px solid ${global_primary_color_100.var}`
     };
     return (
       <tr
-        {...props}
         ref={trRef as React.LegacyRef<HTMLTableRowElement>}
         className={css(className, isOddRow ? 'odd-row-class' : 'even-row-class', 'custom-static-class')}
         style={isOddRow ? customStyle : {}}