+++ packages/react-styles/CHANGELOG.md
@@ -3,6 +3,28 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# 5.2.0-prerelease.2 (2023-10-10)
+
+**Note:** Version bump only for package @patternfly/react-styles
+
+# 5.2.0-prerelease.1 (2023-10-09)
+
+**Note:** Version bump only for package @patternfly/react-styles
+
+# [5.2.0-prerelease.0](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-styles@5.1.1...@patternfly/react-styles@5.2.0-prerelease.0) (2023-10-05)
+
+**Note:** Version bump only for package @patternfly/react-styles
+
+## [5.1.1](https://github.com/patternfly/patternfly-react/compare/@patternfly/react-styles@5.1.1-prerelease.8...@patternfly/react-styles@5.1.1) (2023-10-05)
+
+**Note:** Version bump only for package @patternfly/react-styles
+
+## 5.1.1-prerelease.8 (2023-10-03)
+
+### Bug Fixes
+
+- whitespace changes to trigger prereleases ([#9702](https://github.com/patternfly/patternfly-react/issues/9702)) ([741c248](https://github.com/patternfly/patternfly-react/commit/741c24825b503e116c77cf304aa3e4e3a9ff7841))
+
 ## 5.1.1-prerelease.7 (2023-09-21)
 
 **Note:** Version bump only for package @patternfly/react-styles